refactor(validators): require integer ids in the AI request bodies
reportId and domainId accepted fractional numbers, preserved from the
pre-Zod code. Row ids are integers throughout the schema, so a
fractional id can only ever be a client bug. Error messages unchanged.

+++ validators/ai/diagnosticsInsightsRequestSchema.ts
@@ -6,7 +6,7 @@ import { z } from 'zod'
  */
 export const diagnosticsInsightsRequestSchema = z.object({
   domainName: z.string().min(1),
-  domainId: z.number().min(1),
+  domainId: z.number().int().min(1),
   startDate: z.string().nullish(),
   endDate: z.string().nullish(),
 })
